Rename GuzzleSiteAuthenticator namespace to SiteConfig
