+++ tests/Wallabag/CoreBundle/SiteConfig/GrabySiteConfigBuilderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Wallabag\CoreBundle\GuzzleSiteAuthenticator;
+namespace Tests\Wallabag\CoreBundle\SiteConfig;
 
 use Graby\SiteConfig\ConfigBuilder;
 use Graby\SiteConfig\SiteConfig as GrabySiteConfig;
@@ -10,8 +10,8 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Tests\Wallabag\CoreBundle\WallabagCoreTestCase;
 use Wallabag\CoreBundle\Entity\User;
-use Wallabag\CoreBundle\GuzzleSiteAuthenticator\GrabySiteConfigBuilder;
 use Wallabag\CoreBundle\Repository\SiteCredentialRepository;
+use Wallabag\CoreBundle\SiteConfig\GrabySiteConfigBuilder;
 
 class GrabySiteConfigBuilderTest extends WallabagCoreTestCase
 {
